Give prometheus access to grafana via ALB
I have to grant prometheus access to grafana via ALB. This means that
we may need to open ports or something of that nature. All containers keep
failing graceful, I believe this to be caused by a failing health check.
I need to understand the cause of the fault.

Two ALB's deployed, one targeting prometheus & the other targeting AM. I
am not ceertain if the AM is what is causing the prometheus to die
since it getting a 503 when it tries poll it as shown on logs.