Set DNS names for ALB's & permit prometheus to discover alertmanager via DNS
This commit permits people the abilityt to access promtheus via private DNS name. In order
to achieve this the value of the dns zone id must be set. Once this is set we then need
to ensure that naming convention meets our particular requires. Each DNS name is set within
the application service file.

The prometheus YAML is also updated to reflect this.