Added a script that funcations as required, however permissions are still a problem so this will be investigated tomorrow morning
Shutdown instance if disk is not attached

Edit Script to check if disk is attached via aws cli. If disk
not attached. Loop until disk attached. Timeout and shutdown if disk
not attached in the specified amount of time.

https://trello.com/c/XaZpAVee/444-prevent-user-data-from-starting-if-an-ebs-volume-is-not-attached

WIP, beware main tf edits

The aws command requires a list of volumes.

Add this and ensured that a variable is not set on the script prior to the
point.

Removed JQ in order to ensure & used the default AWS query featurey