# This is a combination of 5 commits. # This is the 1st commit message:
Added a script that funcations as required, however permissions are still a problem so this will be investigated tomorrow morning

# This is the commit message #2:

Shutdown instance if disk is not attached

Edit Script to check if disk is attached via aws cli. If disk
not attached. Loop until disk attached. Timeout and shutdown if disk
not attached in the specified amount of time.

https://trello.com/c/XaZpAVee/444-prevent-user-data-from-starting-if-an-ebs-volume-is-not-attached

# This is the commit message #3:

WIP, beware main tf edits

# This is the commit message #4:

The aws command requires a list of volumes.

Add this and ensured that a variable is not set on the script prior to the
point.

# This is the commit message #5:

Removed JQ in order to ensure & used the default AWS query featurey