Ensure that an EC2 instance will die if not attached to an EBS volume
https://trello.com/c/XaZpAVee/444-prevent-user-data-from-starting-if-an-ebs-volume-is-not-attached

This stops an instance from starting if it cannot attach to an EBS volume. This will mean that if a build fails it will do it in a way that we will notice and that we will not lose metrics. This fixes an issue whereby an instance would fail to attach to a volume because that volume had an old instance still attached.