Pagerduty Configuration
This commit is a straw man proposal for managing Pagerduty with
terrafrom. It is not intended to be merged into master but to
illustrate that it is possible and the it may be something we want to
do.

Issues with this commit are:
1) it is unrelated to the aws infrastructure and so should live in its
own repository
2) does not contain sufficient details about the state file in use
3) does not manage secrets

The issues are not complex and are solvable but I thought it would be
useful to discuss if we want to do this before going ahead.

Currently the state file is in
`s3://gds-tooling-monitoring-ejrowleytest-stack`

to run:

`TF_VAR_pagerduty_token=$(reng-pass pagerduty/user_api_key)  TF_VAR_oncall_name="Joe
Bloggs" TF_VAR_oncall_phonenumber=77665544433 aws-vault exec gds-tech-ops
-- terraform plan -var-file=../../../stacks/example.tfvars`