Build enclave network within Verify Perf
We are deploying Prometheus into Verify Perf, as part of this work we
need to create a subnet from which Prometheus can ran.

The enclave needs to be deployable to multiple verify environments in
future so it was appropriate to create a module which can be customised
per environment.

There is a enclave/network module which describes the resources and
a enclave/verify-perf-a project which instantiates the resources with
perf-a environment configuration. It is then trivial to deploy into new
verify environments by defining an new enclave/verify-{env} project.